backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
SQLAlchemy==2.0.36
python-dotenv==1.0.1
psycopg2-binary==2.9.10
pydantic==2.9.2
passlib==1.7.4
bcrypt==3.2.0
PyJWT==2.4.0
python-multipart==0.0.9
requests==2.32.3